% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{assert_avg_waiting_time_max}
\alias{assert_avg_waiting_time_max}
\title{Assert that average waiting time < max_val}
\usage{
assert_avg_waiting_time_max(envs, max_val)
}
\arguments{
\item{envs}{a list of \code{envs} as produced by \code{run_instance}}

\item{max_val}{the max waiting time duration}
}
\description{
Assert that average waiting time < max_val
}

