% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{constraints_evaluator}
\alias{constraints_evaluator}
\title{Evaluator for the constraint functions}
\usage{
constraints_evaluator(envs, constraints)
}
\arguments{
\item{envs}{a list of \code{envs} as produced by \code{run_instance}}

\item{constraints}{a list of constraint functions}
}
\description{
Evaluator for the constraint functions
}

