% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{grid_optim}
\alias{grid_optim}
\title{A simmer grid optimizer}
\usage{
grid_optim(sim_expr, objective = c("min", "max"), ...)
}
\arguments{
\item{sim_expr}{a function which runs a simmer env and returns a list with at least the keys \code{objective} and \code{constraints}}

\item{objective}{the type of objective to focus on, for now only \code{minimize} and \code{maximize}}

\item{...}{the named arguments and related vectors to optimize over, e.g. \code{var1=1:5, var2=3:5}}
}
\value{
the optimal combination of the variable possibilities supplied in \code{...}
}
\description{
A simmer grid optimizer
}

