% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{method_results}
\alias{method_results}
\title{A helper function to return the results of an optimization method to the optimization framework}
\usage{
method_results(method, objective_value, constraints_satisfied, params,
  envs = NULL, extra_info = list())
}
\arguments{
\item{method}{the name of the optimization function (string)}

\item{objective_value}{the value of the objective}

\item{constraints_satisfied}{boolean indicating whether or not all constraints were satisfied}

\item{params}{the found parameters}

\item{envs}{a copy of the generated envs (optional)}

\item{extra_info}{a list of extra information (optional)}
}
\description{
A helper function to return the results of an optimization method to the optimization framework
}
