% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{msr_arrivals_finished}
\alias{msr_arrivals_finished}
\title{Measure the number of finished arrivals}
\usage{
msr_arrivals_finished(envs, agg = mean)
}
\arguments{
\item{envs}{a list of \code{envs} as produced by \code{run_instance}}

\item{agg}{the method of aggregation of per replication results}
}
\description{
Measure the number of finished arrivals
}

