% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{msr_resource_capacity}
\alias{msr_resource_capacity}
\title{Measure the capacity of a resource type}
\usage{
msr_resource_capacity(envs, name, agg = mean)
}
\arguments{
\item{envs}{a list of \code{envs} as produced by \code{run_instance}}

\item{name}{the name of the resource}

\item{agg}{the method of aggregation of per replication results}
}
\description{
Measure the capacity of a resource type
}

