% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{msr_resource_utilization}
\alias{msr_resource_utilization}
\title{Measure the utilization of a resource type}
\usage{
msr_resource_utilization(envs, name, agg = mean)
}
\arguments{
\item{envs}{a list of \code{envs} as produced by \code{run_instance}}

\item{name}{the name of the resource}

\item{agg}{the method of aggregation of per replication results}
}
\description{
Measure the utilization of a resource type
}

