% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{msr_runtime}
\alias{msr_runtime}
\title{Measure the runtime of the model}
\usage{
msr_runtime(envs, agg = mean)
}
\arguments{
\item{envs}{a list of \code{envs} as produced by \code{run_instance}}

\item{agg}{the method of aggregation of per replication results}
}
\description{
Measure the runtime of the model
}
