% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{objective_evaluator}
\alias{objective_evaluator}
\title{Evaluator for the objective function}
\usage{
objective_evaluator(envs, objective)
}
\arguments{
\item{envs}{a list of \code{envs} as produced by \code{run_instance}}

\item{objective}{an objective function}
}
\description{
Evaluator for the objective function
}

