% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{opt_func}
\alias{opt_func}
\title{Value function generator}
\usage{
opt_func(params)
}
\arguments{
\item{...}{the options passed}
}
\value{
a function which supplies the variables to test
}
\description{
Internal usage
}

