% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{optim_control}
\alias{optim_control}
\title{Control object to configure the optimization procedure}
\usage{
optim_control(run_args = list(until = 1000), replications = 1,
  parallel = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{run_args}{the run arguments, a list with the following arguments: \code{until}}

\item{replications}{the number of replications}

\item{parallel}{whether or not replications should be run in parallel (leveraging \code{mclapply})}

\item{...}{extra named arguments added to the control object}
}
\description{
Control object to configure the optimization procedure
}

