% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs.R
\name{optim_results}
\alias{optim_results}
\title{Function to pass results of simmer evaluation back to the optimization framework}
\usage{
optim_results(objective, constraints = list(), envs = NULL)
}
\arguments{
\item{objective}{the value of the objective}

\item{constraints}{a list with named objectives containing only TRUE and FALSE values}

\item{envs}{the simmer env (OPTIONAL)}
}
\description{
Function to pass results of simmer evaluation back to the optimization framework
}

