% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{par_continuous}
\alias{par_continuous}
\title{A parameter vector of type continuous}
\usage{
par_continuous(vec)
}
\arguments{
\item{vec}{the original vector}
}
\description{
A parameter vector of type continuous
}
