% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{results}
\alias{results}
\title{Show the results of an optimization procedure}
\usage{
results(optim_obj)
}
\arguments{
\item{optim_obj}{the optimization object}
}
\description{
Show the results of an optimization procedure
}

