% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{run_instance}
\alias{run_instance}
\title{Instantiate a simmer expression}
\usage{
run_instance(model, rep = 1, ...)
}
\arguments{
\item{model}{the simmer model}

\item{rep}{the number of replications}

\item{...}{named parameters to be passed to the simmer expression and accessbile through the \code{.opt} variable}
}
\description{
Instantiate a simmer expression
}

