% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_optimized}
\alias{run_optimized}
\title{(re)run a simmer expression using the optimized parameter list}
\usage{
run_optimized(optim_obj)
}
\arguments{
\item{optim_obj}{the optimization object}
}
\description{
For this to work an \code{envs} object has to be returned with the \code{optim_results}
}
\seealso{
results
}

