% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{simmer_optim}
\alias{simmer_optim}
\title{Main entry point for the simmer optimization framework}
\usage{
simmer_optim(model, method, direction = "max",
  objective = msr_arrivals_finished, constraints, params = list(),
  control = optim_control(), ...)
}
\arguments{
\item{model}{the simmer model encapsulated in a function}

\item{method}{the method to be used (e.g. \code{grid_optim})}

\item{direction}{optimization direction (\code{max} or \code{min})}

\item{objective}{the objective function}

\item{constraints}{a list of constraint functions}

\item{params}{a list of parameters to optimize over}

\item{control}{a control list created by a call to \code{optim_control()}}

\item{...}{extra parameters that are passed on to the optimization procedure}
}
\description{
Main entry point for the simmer optimization framework
}
