% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{with_args}
\alias{with_args}
\title{Helper function to run objective / constraint functions with specified arguments}
\usage{
with_args(f, ...)
}
\arguments{
\item{f}{the objective / constraint function}

\item{...}{a list of named arguments which will be used in the call to \code{f}}
}
\description{
Helper function to run objective / constraint functions with specified arguments
}

