% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_optim.R
\name{de_optim}
\alias{de_optim}
\title{A simmer differential evoltion optimizer}
\usage{
de_optim(model, direction = c("min", "max"), objective, constraints,
  params, control, big_m = 1e+06)
}
\arguments{
\item{model}{the simmer model encapsulated in a function}

\item{direction}{optimization direction (\code{max} or \code{min})}

\item{objective}{the objective function}

\item{constraints}{a list of constraint functions}

\item{params}{a list of parameters to optimize over}

\item{control}{a control list created by a call to \code{optim_control()}}

\item{big_m}{a penalty value for solutions with non-satisfied constraints}
}
\description{
Implements the functionality of the \code{DEoptim} package.
}
