% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_optim.R
\name{grid_optim}
\alias{grid_optim}
\title{A simmer grid optimizer}
\usage{
grid_optim(model, direction = c("min", "max"), objective, constraints,
  rep = 1, ...)
}
\arguments{
\item{model}{a function which runs a simmer model}

\item{direction}{optmization direction to focus on, for now only \code{min} and \code{max}}

\item{objective}{the objective measure}

\item{constraints}{a list of constraint functions}

\item{...}{the named arguments and related vectors to optimize over, e.g. \code{var1=1:5, var2=3:5}}
}
\value{
the optimal combination of the variable possibilities supplied in \code{...}
}
\description{
A simmer grid optimizer
}

