% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_optim.R
\name{grid_optim}
\alias{grid_optim}
\title{A simmer grid optimizer}
\usage{
grid_optim(model, direction = c("min", "max"), objective, constraints, params,
  control)
}
\arguments{
\item{model}{the simmer model encapsulated in a function}

\item{direction}{optimization direction (\code{max} or \code{min})}

\item{objective}{the objective function}

\item{constraints}{a list of constraint functions}

\item{params}{a list of parameters to optimize over}

\item{control}{a control list created by a call to \code{optim_control()}}
}
\description{
Executes an exhaustive search over the solution space
}
