% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs.R
\name{opt_func}
\alias{opt_func}
\title{Value function generator}
\usage{
opt_func(...)
}
\arguments{
\item{search_grid}{the grid to search over}

\item{index}{the requested index}
}
\value{
a function which supplies the variables to test
}
\description{
Internal usage
}

