% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{run_instance}
\alias{run_instance}
\title{Instantiate a simmer expression}
\usage{
run_instance(sim_expr, ...)
}
\arguments{
\item{sim_expr}{the simmer expression}

\item{...}{named parameters to be passed to the simmer expression}
}
\description{
Instantiate a simmer expression
}

