% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{run_instance}
\alias{run_instance}
\title{Run n simmer models}
\usage{
run_instance(model, control, params)
}
\arguments{
\item{model}{the simmer model}

\item{control}{the \code{optim_control} object}

\item{params}{a list of named parameters to be passed to the simmer expression and accessbile for the model through the \code{.opt} variable}
}
\description{
Run n simmer models
}
